add day3 working solution
code allows to get the right solution, but it's not very efficient
as it looks for every single coordinate to check if it is in the
array of previous coordinates, algorithm efficiencies can be to
use "High Water Marks" to avoid check the array if HWM is exceeded
(as it will be a new house)... Is there a way to know if we have
already been in a cartesian point without keeping state of the previous
points? if it just was a single dimension, that would be a simple range
of the points already visited, can this be extrapolated somehow to
cartesian points?